refactor: 컨설팅 게시글 컴포넌트 폴더 구조 수정
