+++ src/features/posts/ui/consulting-post-form/consulting-post-form-operation-form.tsx
@@ -5,8 +5,8 @@ import { format } from 'date-fns';
 import { useFormContext } from 'react-hook-form';
 import { cn } from '@/lib/utils';
 import { useState } from 'react';
-import { CONSULTING_POST_FORM_FIELD_NAME } from '../constants/consulting-post-form-field-name';
-import useShowDeleteAllOperations from '../hooks/use-show-delete-all-operations';
+import { CONSULTING_POST_FORM_FIELD_NAME } from '../../constants/consulting-post-form-field-name';
+import useShowDeleteAllOperations from '../../hooks/use-show-delete-all-operations';
 
 export default function ConsultingPostFormOperationForm() {
   const { setValue, getValues } = useFormContext();
